feat: [Multi] 멀티 모드 예외 처리

See merge request s11-webmobile1-sub2/S11P12E202!11

+++ backend/multi/src/main/java/com/scf/multi/application/MultiGameService.java
@@ -5,6 +5,8 @@ import com.scf.multi.domain.dto.Problem;
 import com.scf.multi.domain.dto.Solved;
 import com.scf.multi.domain.model.MultiGameRoom;
 import com.scf.multi.domain.repository.MultiGameRepository;
+import com.scf.multi.global.error.ErrorCode;
+import com.scf.multi.global.error.exception.BusinessException;
 import java.util.List;
 import java.util.Map;
 import java.util.UUID;
@@ -23,7 +25,14 @@ public class MultiGameService {
     }
 
     public MultiGameRoom findOneById(String roomId) {
-        return multiGameRepository.findOneById(roomId);
+
+        MultiGameRoom room = multiGameRepository.findOneById(roomId);
+
+        if (room == null) {
+            throw new BusinessException(roomId, "roomId", ErrorCode.ROOM_NOT_FOUND);
+        }
+
+        return room;
     }
 
     public String createRoom(Long userId) {
@@ -42,14 +51,40 @@ public class MultiGameService {
     }
 
     public void deleteRoom(String roomId) {
+
+        MultiGameRoom room = multiGameRepository.findOneById(roomId);
+
+        if(room == null) {
+            throw new BusinessException(roomId, "roomId", ErrorCode.ROOM_NOT_FOUND);
+        }
+
         multiGameRepository.deleteRoom(roomId);
     }
 
     public void joinRoom(String roomId, JoinRoomDTO joinRoomDTO) {
+
+        MultiGameRoom room = findOneById(roomId);
+
+        if (room.getIsStart()) {
+            throw new BusinessException(roomId, "roomId", ErrorCode.GAME_ALREADY_STARTED);
+        }
+
+        if (room.getPlayers().size() >= room.getMaxPlayer()) {
+            throw new BusinessException(room.getMaxPlayer(), "room max player",
+                ErrorCode.MAX_PLAYERS_EXCEEDED);
+        }
+
         multiGameRepository.joinRoom(roomId, joinRoomDTO);
     }
 
     public void exitRoom(String roomId, Long userId) {
+
+        MultiGameRoom room = multiGameRepository.findOneById(roomId);
+
+        if (room == null) {
+            throw new BusinessException(roomId, "roomId", ErrorCode.ROOM_NOT_FOUND);
+        }
+
         multiGameRepository.exitRoom(roomId, userId);
     }
 
@@ -58,6 +93,10 @@ public class MultiGameService {
         MultiGameRoom room = multiGameRepository.findOneById(roomId);
         List<Problem> problems = room.getProblems();
         Problem problem = problems.get(room.getRound());
+
+        if (problem == null) {
+            throw new BusinessException(null, "problem", ErrorCode.PROBLEM_NOT_FOUND);
+        }
 
         // 문제의 정답 가져오기
         Map<Integer, Integer> answer = problem.getAnswer();
@@ -91,7 +130,15 @@ public class MultiGameService {
 
         MultiGameRoom room = multiGameRepository.findOneById(roomId);
 
+        if (room == null) {
+            throw new BusinessException(roomId, "roomId", ErrorCode.ROOM_NOT_FOUND);
+        }
+
         List<Problem> problems = problemService.getProblems();
+
+        if (problems == null || problems.isEmpty()) {
+            throw new BusinessException(null, "problems", ErrorCode.PROBLEM_NOT_FOUND);
+        }
 
         room.gameStart(problems);
 
@@ -99,6 +146,10 @@ public class MultiGameService {
     }
 
     private int calculateScore(int submitTime) { // TODO: 협의 후 수정
+
+        if(submitTime > 30) {
+            throw new BusinessException(submitTime, "submitTime", ErrorCode.SUBMIT_TIME_EXCEEDED);
+        }
 
         if (submitTime < 3) {
             return 500;
